fix fractional resoure handling for batch

+++ metaflow/plugins/aws/aws_utils.py
@@ -79,7 +79,11 @@ def compute_resource_attributes(decos, compute_deco, resource_defaults):
                     continue
                 if my_val is not None and v is not None:
                     try:
-                        result[k] = str(max(int(my_val or 0), int(v or 0)))
+                        # Use Decimals to compare and convert to string here so
+                        # that numbers that can't be exactly represented as
+                        # floats (e.g. 0.8) still look "nice". We don't care
+                        # about precision more that .001 for resources anyway.
+                        result[k] = str(max(float(my_val or 0), float(v or 0)))
                     except ValueError:
                         # Here, we don't have ints so we compare the value and raise
                         # an exception if not equal
